pyproject.toml
[project]
name = "grid-data"
version = "0.0.1"
dependencies = [
  "dagster",
  "dagster-dbt",
  "dagster-webserver",
  "dbt-duckdb",
  "duckdb",
  "pandas",
  "pyarrow",
  "gridstatus",
  "loguru",
  "python-dotenv"
]
readme = "README.md"

[project.optional-dependencies]
dev = [
  "pre-commit",
  "mypy",
  "black",
  "ruff",
  "pip-tools"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["grid_data"]